Refactored to a cmake-managed header-only library
- reference: https://dominikberner.ch/cmake-interface-lib/
- Top-level CMakeLists.txt installs all callable_traits headers into
  system include folder
- move all headers into callable_traits/include/callable_traits for
  more standardized organization and proper namespacing when installed
  into system include directory
- re-organize test folder by moving all invalid test files to
  test/invalid_test
- updated github workflow ci.yml to reflect changes
