+++ .github/workflows/ci.yml
@@ -22,7 +22,7 @@ jobs:
       run: sudo apt-get update && sudo apt-get install -y cmake g++ ninja-build
 
     - name: Configure CMake
-      run: cmake -B build -G Ninja -DCMAKE_BUILD_TYPE=Release
+      run: cmake -B build -G Ninja -DCMAKE_BUILD_TYPE=Release -DCALLABLE_TRAITS_BUILD_TESTS=ON
 
     - name: Build
       run: cmake --build build
