+++ CMakeLists.txt
@@ -1,13 +1,66 @@
 cmake_minimum_required(VERSION 3.20) # Ensure C++23 support
-project(adapter_factory)
+project("callable_traits" VERSION 1.0.0
+                            DESCRIPTION "A header-only C++ library that provides tools for callable objects."
+                            HOMEPAGE_URL "https://github.com/ndof-opensource/callable_traits.git")
 
 set(CMAKE_CXX_STANDARD 23)
 set(CMAKE_CXX_STANDARD_REQUIRED ON)
 set(CMAKE_CXX_EXTENSIONS OFF)
 
-add_library(callable_traits INTERFACE)
-target_include_directories(callable_traits INTERFACE ${CMAKE_CURRENT_SOURCE_DIR}/include)
+# Option to enable/disable tests
+option(CALLABLE_TRAITS_BUILD_TESTS "Build the test suite" OFF)
 
-# Enable testing and add test directory
-enable_testing()
-add_subdirectory(test)
+include(GNUInstallDirs)
+
+add_library(${PROJECT_NAME} INTERFACE)
+target_include_directories(
+    callable_traits
+    INTERFACE $<BUILD_INTERFACE:${CMAKE_CURRENT_SOURCE_DIR}/include>
+              $<INSTALL_INTERFACE:${CMAKE_INSTALL_INCLUDEDIR}>
+)
+
+target_compile_features(${PROJECT_NAME} INTERFACE cxx_std_23)
+
+# Only enable testing and add test directory if tests are enabled
+if(CALLABLE_TRAITS_BUILD_TESTS)
+    enable_testing()
+    add_subdirectory(test)
+endif()
+
+install(
+    TARGETS ${PROJECT_NAME}
+    EXPORT ${PROJECT_NAME}_Targets
+    ARCHIVE DESTINATION ${CMAKE_INSTALL_LIBDIR}
+    LIBRARY DESTINATION ${CMAKE_INSTALL_LIBDIR}
+    RUNTIME DESTINATION ${CMAKE_INSTALL_BINDIR}
+)
+
+include(CMakePackageConfigHelpers)
+
+write_basic_package_version_file(
+    "${PROJECT_NAME}ConfigVersion.cmake"
+    VERSION ${PROJECT_VERSION}
+    COMPATIBILITY SameMajorVersion
+)
+
+configure_package_config_file(
+    "${PROJECT_SOURCE_DIR}/cmake/${PROJECT_NAME}_config.cmake.in"
+    "${PROJECT_BINARY_DIR}/${PROJECT_NAME}Config.cmake"
+    INSTALL_DESTINATION
+    ${CMAKE_INSTALL_DATAROOTDIR}/${PROJECT_NAME}/cmake
+)
+
+install(
+    EXPORT ${PROJECT_NAME}_Targets
+    FILE ${PROJECT_NAME}Targets.cmake
+    NAMESPACE ${PROJECT_NAME}::
+    DESTINATION ${CMAKE_INSTALL_DATAROOTDIR}/${PROJECT_NAME}/cmake
+)
+
+install(
+    FILES "${PROJECT_BINARY_DIR}/${PROJECT_NAME}Config.cmake"
+          "${PROJECT_BINARY_DIR}/${PROJECT_NAME}ConfigVersion.cmake"
+    DESTINATION ${CMAKE_INSTALL_DATAROOTDIR}/${PROJECT_NAME}/cmake
+)
+
+install(DIRECTORY ${PROJECT_SOURCE_DIR}/include/callable_traits DESTINATION include)
