+++ test/invalid_test/compile_fail_check.cmake
@@ -1,4 +1,4 @@
-# test/compile_fail_check.cmake
+# test/invalid_test/compile_fail_check.cmake
 
 cmake_minimum_required(VERSION 3.20)
 message(STATUS "Checking that invalid concept usage fails to compile...")
@@ -6,10 +6,9 @@ message(STATUS "Checking that invalid concept usage fails to compile...")
 # Configure try_compile
 try_compile(COMPILE_SUCCEEDED
   "${CMAKE_BINARY_DIR}/compile_fail"
-  "${CMAKE_SOURCE_DIR}/test"
+  "${CMAKE_CURRENT_LIST_DIR}"
   invalid_concepts_test.cpp
   CMAKE_FLAGS
-    "-DINCLUDE_DIRECTORIES=${CMAKE_SOURCE_DIR}/include"
     "-DCMAKE_CXX_STANDARD=23"
     "-DCMAKE_CXX_STANDARD_REQUIRED=ON"
     "-DCMAKE_CXX_EXTENSIONS=OFF"
@@ -21,5 +20,4 @@ if(COMPILE_SUCCEEDED)
     "Expected invalid_concepts_test.cpp to fail compilation, but it succeeded.\nOutput:\n${OUTPUT}")
 else()
   message(STATUS "✅ invalid_concepts_test.cpp correctly failed to compile.")
-endif()
-
+endif()
