+++ test/test_callable_concepts.cpp
@@ -1,6 +1,6 @@
 #include <gtest/gtest.h>
 #include <functional>
-#include "callable_concepts.hpp"
+#include <callable_traits/callable_concepts.hpp>
 
 using namespace ndof;
